Updated FTXUI version from 4.0.0 to 5.0.0
Currently, CMake fetches FTXUI's content from the 4.0.0 version of FTXUI.

Because of this, some of the examples provided in the main github repository of FTXUI won't work. (For example: examples/component/window.cpp will not work).

I've updated the GIT_TAG version from "v4.0.0" to "v5.0.0" to ensure that all examples work.